Move GoogleSignInViewController to its own file
