+++ Food/Supporting/GoogleSignInManager.swift
@@ -10,7 +10,6 @@ import GoogleSignIn
 
 class GoogleSignInManager: NSObject, GIDSignInDelegate, ObservableObject {
 
-    var silly: Bool = true
     var isSignedIn: Bool {
         GIDSignIn.sharedInstance()?.hasPreviousSignIn() ?? false
     }
@@ -21,12 +20,10 @@ class GoogleSignInManager: NSObject, GIDSignInDelegate, ObservableObject {
             return
         }
 
-        print("signed in (:")
         objectWillChange.send()
     }
 
     func sign(_ signIn: GIDSignIn!, didDisconnectWith user: GIDGoogleUser!, withError error: Error!) {
-        print("signed out...")
         objectWillChange.send()
     }
